pyproject.toml
[tool.poetry]
name = "strokepred"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.10 || 3.11.11"
pandas = "<2.2.0"
kagglehub = "^0.3.3"
pycaret = "^3.3.2"
kaleido = "0.2.1"
matplotlib = "<3.8.0"
imbalanced-learn = "^0.12.4"


[tool.poetry.group.notebook.dependencies]
jupyter = "^1.1.1"
ydata-profiling = "^4.12.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
pylint = "^3.3.1"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-with-pdf = "^0.9.3"
mkdocs-terminal = "^4.6.0"
mkdocs-material = "^9.5.44"


[tool.poetry.group.airflow.dependencies]
apache-airflow = "*"
apache-airflow-providers-docker = "*"
gspread = "^6.1.4"
pymsteams = "^0.2.3"
apache-airflow-providers-microsoft-azure = "^12.0.0"


[tool.poetry.group.api.dependencies]
fastapi = {extras = ["standard"], version = "^0.115.6"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
